#134 master adaptation
