+++ restservice/src/main/java/org/itevents/dao/mybatis/builder/EventSqlBuilder.java
@@ -12,17 +12,23 @@ public class EventSqlBuilder {
 
     public String addEventTechnology(final Event event) {
         List<Technology> technologies = event.getTechnologies();
-        StringBuilder sql = new StringBuilder();
-        if (!CollectionUtils.isEmpty(technologies)) {
-            sql.append("INSERT INTO event_technology (event_id, technology_id) VALUES ");
-            Iterator<Technology> iterator = technologies.iterator();
-            while (iterator.hasNext()) {
-                sql.append("(").append(event.getId()).append(", ").append(iterator.next().getId()).append(")");
-                if (iterator.hasNext()) {
-                    sql.append(", ");
-                }
-            }
+        if (CollectionUtils.isEmpty(technologies)) {
+            return "";
+        } else {
+            return makeSql(event, technologies);
         }
+    }
+
+    private String makeSql(Event event, List<Technology> technologies) {
+        StringBuilder sql = new StringBuilder();
+        sql.append("INSERT INTO event_technology (event_id, technology_id) VALUES ");
+        for (Technology technology : technologies) {
+            sql.append("(")
+                    .append(event.getId()).append(", ")
+                    .append(technology.getId()).append("), ");
+        }
+        int lastCommaPositionFromEnd = 2;
+        sql.delete(sql.length() - lastCommaPositionFromEnd, sql.length());
         return sql.toString();
     }
